Add potential fixes for #54.
This string will cause an error to be thrown when written:
^[[1;54r^[[H^[[11M^[[0;7m^O

It breaks down to these calls:
setScrollRegion([1, 54]);
cursorPos([1, 1]);
deleteLines([11]);
charAttributes([0, 7]);
setgLevel(0); // shift-in, 0x0f

^[[1;54r sets the scrollBottom to 54 which is bigger than the height of the
terminal (25 in the original issue). This may be due to a mismatch between what
the program thinks is the terminal height and what the terminal knows is its
height. In other words, it's the programs fault. We can add a failsafe for this
in multiple places.